Refactor FallDetector and TFImageDetection code
